Moved abstract/generic/helper classes into classes directory
