+++ components/RelatedPosts.php
@@ -9,6 +9,8 @@ use GinoPane\BlogTaxonomy\Plugin;
 use GinoPane\BlogTaxonomy\Models\Tag;
 use October\Rain\Database\Collection;
 use GinoPane\BlogTaxonomy\Models\ModelAbstract;
+use GinoPane\BlogTaxonomy\Classes\ComponentAbstract;
+use GinoPane\BlogTaxonomy\Classes\TranslateArrayTrait;
 
 /**
  * Class RelatedPosts
