+++ components/SeriesList.php
@@ -5,6 +5,8 @@ namespace GinoPane\BlogTaxonomy\Components;
 use Cms\Classes\Page;
 use GinoPane\BlogTaxonomy\Plugin;
 use GinoPane\BlogTaxonomy\Models\Series;
+use GinoPane\BlogTaxonomy\Classes\ComponentAbstract;
+use GinoPane\BlogTaxonomy\Classes\TranslateArrayTrait;
 
 /**
  * Class SeriesList
