+++ components/TagList.php
@@ -6,6 +6,8 @@ use Cms\Classes\Page;
 use GinoPane\BlogTaxonomy\Plugin;
 use GinoPane\BlogTaxonomy\Models\Tag;
 use October\Rain\Database\Collection;
+use GinoPane\BlogTaxonomy\Classes\ComponentAbstract;
+use GinoPane\BlogTaxonomy\Classes\TranslateArrayTrait;
 
 /**
  * Class TagList
