+++ components/TagPosts.php
@@ -5,6 +5,7 @@ namespace GinoPane\BlogTaxonomy\Components;
 use Rainlab\Blog\Models\Post;
 use GinoPane\BlogTaxonomy\Plugin;
 use GinoPane\BlogTaxonomy\Models\Tag;
+use GinoPane\BlogTaxonomy\Classes\PostListAbstract;
 
 /**
  * Class TagPosts
